Cleanup, restructuring
Removed utils.rs for now, and moved the containing code to explosive.rs. Also formatting
